feat: refactor, move some components
