--- src/app/[locale]/(logged-in)/organisation/profile/components/UserInvite/UserInvite.tsx
@@ -1,9 +1,0 @@
-import InviteUser from "@/modules/InviteUser";
-
-interface UserInviteProps {
-  organisationId?: number;
-}
-
-export default function UserInvite({ organisationId }: UserInviteProps) {
-  <InviteUser organisationId={organisationId} />;
-}
