--- src/app/[locale]/(logged-in)/organisation/profile/components/UserInvite/index.ts
@@ -1,3 +1,0 @@
-import UserInvite from "./UserInvite";
-
-export default UserInvite;
