--- src/components/UserModal/index.ts
@@ -1,5 +1,0 @@
-import UserModal from "./UserModal";
-
-export default UserModal;
-
-export type * from "./UserModal";
